LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/window-open-self-expected.txt



FAIL No replace before load, triggered by window.open(_self) on an iframe assert_equals: history.length must increase expected 18 but got 17
